feature(feature/162727466/users-should-be-able-to-reset-password): Refactored User Middleware Abstracted base url [Fixes #162727466]
