+++ src/helpers/basePath.js
@@ -1,0 +1,6 @@
+const basePath = (req) => {
+  const path = `${req.protocol}://${req.headers.host}`;
+  return path;
+};
+
+export default basePath;
